Implemented logic to check if Skill already exists before creating Implemented taskcreate logic(adds  task to skill if it exists)

+++ Controllers/SkillController.cs
@@ -1,6 +1,7 @@
 ﻿using Azure;
 using JWT_Login_Authorization_DotNet.Interfaces;
 using JWT_Login_Authorization_DotNet.Models;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using System.Linq;
@@ -33,14 +34,25 @@ namespace JWT_Login_Authorization_DotNet.Controllers
         [HttpPost]
         public async Task<IActionResult> PostAsync([FromQuery] SkillDTO skillDTO)
         {
-            if (!ModelState.IsValid)
+            try
             {
-                return BadRequest(ModelState);
-            }
-            Skill azureSkill = _skillService.MapSkill(skillDTO).Result;
+                if (await _skillService.CheckIfSkillExistsAsync(skillDTO.Name) != false)
+                {
+                    return StatusCode(403, "Cannot create skill as it already exists");
+                }
+                if (!ModelState.IsValid)
+                {
+                    return BadRequest(ModelState);
+                }
+                Skill azureSkill = _skillService.MapSkill(skillDTO).Result;
 
-            await _skillService.CreateSkillAsync(azureSkill);
-            return Ok("You have sucessully created a task: " + azureSkill.Name + " with the following ID: " + azureSkill.Id);
+                await _skillService.CreateSkillAsync(azureSkill);
+                return Ok("You have sucessully created a task: " + azureSkill.Name + " with the following ID: " + azureSkill.Id);
+            }
+            catch (RequestFailedException ex)
+            {
+                return StatusCode(ex.Status, ex.Message);
+            }
         }
 
         [HttpGet("Skills")]
@@ -60,15 +72,37 @@ namespace JWT_Login_Authorization_DotNet.Controllers
             {
                 await _skillService.DeleteSkillAsync(SkillId, SkillName);
                 List<Models.Task> tasks = await _taskTableStorageService.GetTasksBySkillName(SkillName);
-                foreach (Models.Task task in tasks)
+                if (tasks.Count > 0)
                 {
-                    await _taskTableStorageService.DeleteTaskAsync(task.Id, SkillName);
+                    foreach (Models.Task task in tasks)
+                    {
+                        await _taskTableStorageService.DeleteTaskAsync(task.Id, SkillName);
+                    }
                 }
                 return Ok("You have sucessufully deleted : " + SkillName + " " + SkillId + " from storage");
             }
             catch (RequestFailedException ex)
             {
                 return StatusCode(ex.Status, ex.Message);
+            }
+        }
+
+        [HttpPut("UpdateTask"), AllowAnonymous]
+        public async Task<IActionResult> UpdateAsync(string skillId, [FromQuery] SkillDTO skillDTO)
+        {
+            if (string.IsNullOrWhiteSpace(skillId))
+            {
+                return BadRequest("SkillID cannot be empty");
+            }
+
+            try
+            {
+                List<Models.Task> tasks = await _taskTableStorageService.GetTasksBySkillName(skillDTO.Name);
+                return Ok(tasks);
+            }
+            catch (RequestFailedException azureEx)
+            {
+                return StatusCode(azureEx.Status, azureEx.Message);
             }
         }
     }
